Moved debits API data fetch logic to App.js
