+++ src/App.js
@@ -5,7 +5,8 @@ import {BrowserRouter as Router, Route} from 'react-router-dom';
 import Home from './components/Home';
 import UserProfile from './components/UserProfile';
 import LogIn from './components/Login';
-import DebitsData from './components/DebitsData';
+import Debits from './components/Debits';
+import axios from 'axios';
 
 class App extends Component {
   constructor() {  // Create and initialize state
@@ -14,7 +15,9 @@ class App extends Component {
       currentUser: {
         userName: 'Joe Smith',
         memberSince: '07/23/96',
-      }
+      },
+      accountBalance:0,
+      debits:[]
     }
   }
 
@@ -26,6 +29,26 @@ class App extends Component {
     this.setState({currentUser: newUser})
   }
 
+  // Make async API call to retrieve data from remote website
+  async componentDidMount() {
+    let linkToAPI = 'https://moj-api.herokuapp.com/debits';  // Link to remote website API
+
+    // Await for promise (completion) returned from API call
+    try {  // Accept success response as array of JSON objects (debits)
+      let response = await axios.get(linkToAPI);
+      console.log(response);  // Print out response
+      // To get data object in the response, need to use "response.data"
+      this.setState({debits: response.data});  // Store received data in state's "debits" object
+    } 
+    catch (error) {  // Print out errors at console when there is an error response
+      if (error.response) {
+        // The request was made, and the server responded with error message and status code.
+        console.log(error.response.data);  // Print out error message (e.g., Not Found)
+        console.log(error.response.status);  // Print out error status code (e.g., 404)
+      }    
+    }
+  }  
+
   // Create Routes and React elements to be rendered using React components
   render() {  
     const HomeComponent = () => (<Home/>);
@@ -33,7 +56,7 @@ class App extends Component {
       <UserProfile userName={this.state.currentUser.userName} memberSince={this.state.currentUser.memberSince}  />
     );
     const LogInComponent = () => (<LogIn user={this.state.currentUser} mockLogIn={this.mockLogIn} />)  // Pass props to "LogIn" component
-    const DebitsComponent = () =>(<DebitsData/>);      
+    const DebitsComponent = () =>(<Debits accountBalance={this.state.accountBalance} debits={this.state.debits}/>);      
     return (
       <Router>
         <div>
