+++ src/components/AccountBalance.js
@@ -5,9 +5,9 @@ import React, {Component} from 'react';
 class AccountBalance extends Component {
   render() {
     return (
-      <div>
+      <h2 className='account-balance'>
         Balance: {this.props.accountBalance}
-      </div>
+      </h2>
     );
   }
 }
